requirements.txt
python-dotenv
openai
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
